refactor(forms): share form attachment row
